requirements.txt
fastapi
uvicorn
pymongo[srv]
passlib[bcrypt]
pydantic
python-dotenv
motor
email-validator
python-jose[cryptography]
bcrypt
requests
PyJWT
transformers
pysentimiento
nemo_toolkit[asr] 
torchaudio 
pydub
python-multipart
numpy==1.26.4